requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
lxml>=4.9.0
pyarrow
pydfs-lineup-optimizer-enhanced
nfl_data_py
lxml>=4.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
plotly>=5.17.0
nbastatpy>=0.1.0
cs2api>=0.3.0
leaguepedia_parser_thomasbarrepitous
flashscore-scraper
scikit-learn
PyYAML
auto-machine-learning
mem0ai